+++ src/components/common/TextInput.vue
@@ -1,23 +1,20 @@
 <template>
   <div class="relative">
     <XMarkIcon
-      v-if="beforeClose"
+      v-if="beforeClose && clearable"
       class="absolute right-2 top-1/2 z-10 h-3 w-3 -translate-y-1/2 cursor-pointer"
       @click="clearInput"
     />
     <input
       v-model="inputValue"
       type="text"
-      class="input input-sm join-item input-bordered w-full"
+      :class="['input input-sm join-item input-bordered w-full', { 'pr-6': clearable }]"
       :placeholder="placeholder || ''"
       :name="name || ''"
       :autocomplete="autocomplete || ''"
-      @touchstart.stop
-      @touchmove.stop
-      @touchend.stop
     />
     <XMarkIcon
-      v-if="!beforeClose"
+      v-if="!beforeClose && clearable"
       class="absolute right-2 top-1/2 z-10 h-3 w-3 -translate-y-1/2 cursor-pointer"
       @click="clearInput"
     />
@@ -32,6 +29,7 @@ defineProps<{
   beforeClose?: boolean
   name?: string
   autocomplete?: string
+  clearable?: boolean
 }>()
 
 const inputValue = defineModel<string>()