+++ src/components/settings/LanguageSelect.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="flex items-center gap-2">
-    {{ $t('language') }}:
+    {{ $t('language') }}
     <select
       class="select select-bordered select-sm w-48"
       v-model="language"
@@ -18,7 +18,7 @@
 </template>
 
 <script setup lang="ts">
-import { LANG } from '@/config'
+import { LANG } from '@/constant'
 import { i18n } from '@/i18n'
 import { language } from '@/store/settings'